Module ntru::encparams

NTRUEncrypt Encryption parameters

This module contains the parameters for NTRUEncrypt. Theese parameters must be used when encrypting, decrypting and generating key pairs. The recomendation is to use the default parameters for each level of security, and not use the deprecated parameters. The recommended parameters are the following:

DEFAULT_PARAMS_112_BITS for 112 bits of security.

DEFAULT_PARAMS_128_BITS for 128 bits of security.

DEFAULT_PARAMS_192_BITS for 192 bits of security.

DEFAULT_PARAMS_256_BITS for 256 bits of security.

Structs

EncParams
A set of parameters for NtruEncrypt

Constants

ALL_PARAM_SETS
All parameter sets, in an array

DEFAULT_PARAMS_112_BITS
The default parameter set for 112 bits of security.

DEFAULT_PARAMS_128_BITS
The default parameter set for 128 bits of security.

DEFAULT_PARAMS_192_BITS
The default parameter set for 192 bits of security.

DEFAULT_PARAMS_256_BITS
The default parameter set for 256 bits of security.

EES1087EP1
An IEEE 1361.1 parameter set that gives 192 bits of security and is optimized for encryption/decryption speed.

EES1087EP2
An IEEE 1361.1 parameter set that gives 256 bits of security and is optimized for key size.

EES1171EP1
An IEEE 1361.1 parameter set that gives 256 bits of security and is a tradeoff between key size and encryption/decryption speed.

EES1499EP1
An IEEE 1361.1 parameter set that gives 256 bits of security and is optimized for encryption/decryption speed.

EES401EP1
An IEEE 1361.1 parameter set that gives 112 bits of security and is optimized for key size.

EES401EP2
A product-form parameter set that gives 112 bits of security.

EES439EP1
DEPRECATED A product-form parameter set that gives 128 bits of security.

EES443EP1
A product-form parameter set that gives 128 bits of security.

EES449EP1
An IEEE 1361.1 parameter set that gives 128 bits of security and is optimized for key size.

EES541EP1
An IEEE 1361.1 parameter set that gives 112 bits of security and is a tradeoff between key size and encryption/decryption speed.

EES587EP1
A product-form parameter set that gives 192 bits of security.

EES593EP1
DEPRECATED A product-form parameter set that gives 192 bits of security.

EES613EP1
An IEEE 1361.1 parameter set that gives 128 bits of security and is a tradeoff between key size and encryption/decryption speed.

EES659EP1
An IEEE 1361.1 parameter set that gives 112 bits of security and is optimized for encryption/decryption speed.

EES677EP1
An IEEE 1361.1 parameter set that gives 192 bits of security and is optimized for key size.

EES743EP1
A product-form parameter set that gives 256 bits of security.

EES761EP1
An IEEE 1361.1 parameter set that gives 128 bits of security and is optimized for encryption/decryption speed.

EES887EP1
An IEEE 1361.1 parameter set that gives 192 bits of security and is a tradeoff between key size and encryption/decryption speed.