Function ntru::generate_multiple_key_pairs

pub fn generate_multiple_key_pairs(params: &EncParams, rand_context: &RandContext, num_pub: usize) -> Result<(PrivateKey, Box<[PublicKey]>), Error>

Key generation with multiple public keys

Generates num_pub NtruEncrypt key pairs. They all share a private key but their public keys differ. The private key decrypts messages encrypted for any of the public keys. Note that when decrypting, the public key of the key pair passed into ntru_decrypt() must match the public key used for encrypting the message. If a deterministic RNG is used, the key pair will be deterministic for a given random seed; otherwise, the key pair will be completely random.