Struct ntru::types::IntPoly

pub struct IntPoly {
    // some fields omitted
}

A polynomial with integer coefficients.

Methods

impl IntPoly

fn new(coeffs: &[i16]) -> IntPoly

Create a new IntPoly

fn rand(n: u16, pow2q: u16, rand_ctx: &RandContext) -> IntPoly

Create a new random IntPoly

fn from_arr(arr: &[u8], n: u16, q: u16) -> IntPoly

Convert array to IntPoly

fn get_coeffs(&self) -> &[i16]

Get the coefficients

fn set_coeffs(&mut self, coeffs: &[i16])

Set the coefficients

fn set_coeff(&mut self, index: usize, value: i16)

Set a coefficient

fn mod_mask(&mut self, mod_mask: u16)

Modifies the IntPoly with the given mask

fn to_arr(&self, params: &EncParams) -> Box<[u8]>

Converts the IntPoly to a byte array using 32 bit arithmetic

fn mult_tern(&self, b: &TernPoly, mod_mask: u16) -> (IntPoly, bool)

General polynomial by ternary polynomial multiplication

Multiplies a IntPoly by a TernPoly. The number of coefficients must be the same for both polynomials. It also returns if the number of coefficients differ or not.

fn add_tern(&self, b: &TernPoly) -> IntPoly

Add a ternary polynomial

Adds a ternary polynomial to the general polynomial. Returns a new general polynomial.

fn mult_prod(&self, b: &ProdPoly, mod_mask: u16) -> (IntPoly, bool)

General polynomial by product-form polynomial multiplication

Multiplies a IntPoly by a ProdPoly. The number of coefficients must be the same for both polynomials. It also returns if the number of coefficients differ or not.

fn mult_priv(&self, b: &PrivPoly, mod_mask: u16) -> (IntPoly, bool)

General polynomial by private polynomial multiplication

Multiplies a IntPoly by a PrivPoly, i.e. a TernPoly or a ProdPoly. The number of coefficients must be the same for both polynomials. It also returns if the number of coefficients differ or not.

fn mult_int(&self, b: &IntPoly, mod_mask: u16) -> (IntPoly, bool)

General polynomial by general polynomial multiplication

Multiplies a IntPoly by another IntPoly, i.e. a TernPoly or a ProdPoly. The number of coefficients must be the same for both polynomials. It also returns if the number of coefficients differ or not.

fn mult_fac(&mut self, factor: i16)

Multiply by factor

fn mod_center(&mut self, modulus: u16)

Calls ntru_mod_center() in this polinomial.

fn mod3(&mut self)

Calls ntru_mod3() in this polinomial.

fn equals_mod(&self, other: &IntPoly, modulus: u16) -> bool

Check if both polynomials are equals given a modulus

fn equals1(&self) -> bool

Check if the IntPoly equals 1