Struct ntru::types::ProdPoly

pub struct ProdPoly {
    // some fields omitted
}

A product-form polynomial, i.e. a polynomial of the form f1*f2+f3 where f1,f2,f3 are very sparsely populated ternary polynomials.

Methods

impl ProdPoly

fn new(n: u16, f1: TernPoly, f2: TernPoly, f3: TernPoly) -> ProdPoly

Creates a new ProdPoly from three TernPolys

fn rand(n: u16, df1: u16, df2: u16, df3_ones: u16, df3_neg_ones: u16, rand_ctx: &RandContext) -> Option<ProdPoly>

Random product-form polynomial

Generates a random product-form polynomial consisting of 3 random ternary polynomials. Parameters:

N: the number of coefficients, must be MAX_DEGREE or less

df1: number of ones and negative ones in the first ternary polynomial

df2: number of ones and negative ones in the second ternary polynomial

df3_ones: number of ones ones in the third ternary polynomial

df3_neg_ones: number of negative ones in the third ternary polynomial

rand_ctx: a random number generator

fn to_int_poly(&self, modulus: u16) -> IntPoly

Returns an IntPoly equivalent to the ProdPoly